Added volume_type field function.
Volume types in ceilometer samples are stored as uuids. This
change adds a way to convert volume type uuids to volume type names.
One the first call to the volume_type field function, the cinder
client is used to gather a list of volume types. This is cached and
usable for every volume resource afterwards. If volume_type is never
used, the cinder client will not be invoked to load volume types.